Change qualified names for connection profiles- identity cards (#1736)
with a web profile now share a fixed connection profile, and
other connection profile names are now prefixed with a hash of
the connection profile object.

Also adds a function to get an idcard with required role and
updates the connection profile service to avoid overwriting
existing connection profiles.

Contributes to #902

Signed-off-by: James Taylor <jamest@uk.ibm.com>

(cherry picked from commit fac9547)