Prevent initial identity cards being deleted (#1997)
It is currently possible to delete every identity card in the playground
and end up with an empty wallet on the login screen. While refreshing
the screen does recreate the initial identity cards created by the
playground automatically, it is a confusing state to get into. This
fix prevents any of the initial identity cards being deleted. It
might be that we want to allow, or even encourage, people to delete
some of those admin cards in the future, however that would currently
leave no way to get them back if required and this seems like a
reasonable way to address the problem for now.

Contributes to #1930

Signed-off-by: James Taylor <jamest@uk.ibm.com>