only allow a single enrollment per issued identity
This pull request addresses
https://github.com/hyperledger/composer/issues/2056

where rather than allowing an issued identity to be enrolled indefinitely it restricts it to a single enrollment and if you lose the crypto material then you cannot automatically re-enroll the user. Basically it will stop the confusion of identity not being registered when a user someone re-enrolls a user they had already bound to a participant and not understand why it doesn’t work as effectively it is a new certificate (admittedly for the same identity) and it would then output a different error message this time from the CA saying cannot enroll.

Signed-off-by: Dave Kelsey <d_kelsey@uk.ibm.com>