Version playground data in browser
Playground no longer prompts users to delete local storage, which
was not optional anyway, and instead stores data keyed by composer
version. This should still ensure that left over data does not cause
problems with the 'Get started' flow, while leaving the possibility
of retreiving old data.

It would be nice for data to persist accross versions, see #1447,
however this is less of an issue if playground is focussed on the
first use experience.

Closes #3873

Signed-off-by: James Taylor <jamest@uk.ibm.com>