Improve Embedded Connector Identity Management
closes #2857

This PR includes the ability to host the embedded connector in a server via the composer-connector-server and drive it from client applications such as CLI or Rest Server. It will probably also work with playground but only if running locally.

Integration tests would be required + docs to make this an officially supported capability so for now, it won’t be documented.

Signed-off-by: Dave Kelsey <d_kelsey@uk.ibm.com>